fix(build): enforce app guardrails and dflow boundaries
