+++ .github/workflows/app-build.yml
@@ -1,0 +1,32 @@
+name: App Build
+
+on:
+  pull_request:
+    types: [opened, synchronize, reopened]
+
+permissions:
+  contents: read
+
+jobs:
+  build-app:
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: app
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Bun
+        uses: oven-sh/setup-bun@v2
+        with:
+          bun-version: "1.2.17"
+
+      - name: Install app dependencies
+        run: bun install --frozen-lockfile
+
+      - name: Lint app
+        run: bun run lint
+
+      - name: Build app
+        run: bun run build
